Making environment utility reusable
